+++ packages/qti-convert-core/src/qti-downgrader/index.ts
@@ -1,2 +1,3 @@
 export * from './convert-qti3-to-qti21';
 export * from './package-converter';
+export * from './qti3-shared-vocabulary-css';